+++ internal/table/client.go
@@ -465,10 +465,10 @@ func (c *client) Close(ctx context.Context) (err error) {
 	c.wgClosed.Wait()
 
 	if len(issues) > 0 {
-		return errors.NewWithIssues(
+		return errors.Error(errors.NewWithIssues(
 			"table client closed with issues",
 			issues...,
-		)
+		))
 	}
 
 	return nil